Set up disambiguation resolution and move old code out
